Extract messages and commands to their own file
