+++ tui/stopwatch/stopwatch.go
@@ -1,8 +1,6 @@
 package stopwatch
 
 import (
-	"time"
-
 	"github.com/charmbracelet/bubbles/key"
 	tea "github.com/charmbracelet/bubbletea"
 )
@@ -17,47 +15,6 @@ func Run() {
 	if _, err := p.Run(); err != nil {
 		panic(err)
 	}
-}
-
-// MESSAGES & COMMANDS
-// -------------------
-
-// A message to start the stopwatch
-type msgStart struct{ t time.Time }
-
-// A message to stop the stopwatch
-type msgStop struct{}
-
-// A message to resume the stopwatch
-type msgResume struct{}
-
-type msgLap struct{}
-
-// A message to reset the stopwatch
-type msgReset struct{}
-
-// A command to start the stopwatch
-func (m *stopwatchModel) Start() tea.Msg {
-	return msgStart{time.Now()}
-}
-
-// A command to stop the stopwatch
-func (m *stopwatchModel) Stop() tea.Msg {
-	return msgStop{}
-}
-
-// A command to resume the stopwatch
-func (m *stopwatchModel) Resume() tea.Msg {
-	return msgResume{}
-}
-
-// A command to reset the stopwatch
-func (m *stopwatchModel) Reset() tea.Msg {
-	return msgReset{}
-}
-
-func (m *stopwatchModel) Lap() tea.Msg {
-	return msgLap{}
 }
 
 // KEYBINDINGS
